requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
